Rewrite chrome.windows.onRemoved and chrome.tabs.onRemoved callbacks
When we close a window or a tab, we need to reset the relevant `windowId`
or `tagId` value to `0`.

We had scattered `chrome.windows.onRemoved` and `chrome.tabs.onRemoved`
callbacks all around `scripts/background.js`.

We organize them all in the same place and remove a lot of duplicate
code.