Remove 'showall' handling in scripts/background.js
When running `scripts/background.js`, we have one place in the code
where we check for each individual option and another place in the code
where we check for `showall`. In both cases, we do the same things if
every option is checked.

There is no point checking specifically for `showall` in the code.
Its a UI convenience, there shouldn't be any special code about it in
`scripts/background.js`. Therefore, we remove it with this PR.

The functionality remains exactly the same.